fix_2 objects to another package
