+++ hw05_shapes/objects.go
@@ -1,0 +1,13 @@
+package main
+
+type Circle struct {
+	r float64
+}
+
+type Rectangle struct {
+	x, y float64
+}
+
+type Triangle struct {
+	x, h float64
+}
